fix:修复微信支付小于1元不能发起支付的bug

+++ app/Plugins/User/src/Controller/AssetController.php
@@ -38,7 +38,7 @@ class AssetController
     public function money_recharge_submit()
     {
         $payment = request()->input('payment');
-        $amount = request()->input('amount');
+        $amount = (float) request()->input('amount');
         $captcha = request()->input('captcha');
         if (!$payment || !$amount || !$captcha) {
             return Json_Api(403, false, ['msg' => '请求参数不足']);
